% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gs}
\alias{Gs}
\title{Gs
The Gs value used in CLiff walk (Description)}
\format{
An object of class \code{numeric} of length 10000.
}
\usage{
data(Gs)
}
\description{
Gs
The Gs value used in CLiff walk (Description)
}
\keyword{datasets}
