% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mafor5.14}
\alias{mafor5.14}
\title{mafor5.14
the matrix for homework 5.14}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 4 columns.
}
\usage{
data(mafor5.14)
}
\description{
mafor5.14
the matrix for homework 5.14
}
\keyword{datasets}
