% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{powerfor6_8}
\alias{powerfor6_8}
\title{powerfor6_8
The power value used in homework 6.8 (Description)}
\format{
An object of class \code{numeric} of length 15.
}
\usage{
data(powerfor6_8)
}
\description{
powerfor6_8
The power value used in homework 6.8 (Description)
}
\keyword{datasets}
