% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Q}
\alias{Q}
\title{Q
The Q value used in CLiff walk (Description)}
\format{
An object of class \code{array} of dimension 4 x 12 x 4.
}
\usage{
data(Q)
}
\description{
Q
The Q value used in CLiff walk (Description)
}
\keyword{datasets}
