% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2.1forQs}
\alias{r2.1forQs}
\title{r2.1forQs
The res r2.1 value used in homework Qs (Description)}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
data(r2.1forQs)
}
\description{
r2.1forQs
The res r2.1 value used in homework Qs (Description)
}
\keyword{datasets}
