% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2forQs}
\alias{r2forQs}
\title{r2forQs
The res r2 value used in homework Qs (Description)}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
data(r2forQs)
}
\description{
r2forQs
The res r2 value used in homework Qs (Description)
}
\keyword{datasets}
