% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r3.1forQs}
\alias{r3.1forQs}
\title{r3.1forQs
The res r3.1 value used in homework Qs (Description)}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
data(r3.1forQs)
}
\description{
r3.1forQs
The res r3.1 value used in homework Qs (Description)
}
\keyword{datasets}
