% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r3forQs}
\alias{r3forQs}
\title{r3forQs
The res r3 value used in homework Qs (Description)}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
data(r3forQs)
}
\description{
r3forQs
The res r3 value used in homework Qs (Description)
}
\keyword{datasets}
