% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r4forQs}
\alias{r4forQs}
\title{r4forQs
The res r4 value used in homework Qs (Description)}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
data(r4forQs)
}
\description{
r4forQs
The res r4 value used in homework Qs (Description)
}
\keyword{datasets}
