% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resfor6_8}
\alias{resfor6_8}
\title{resfor6_8
The res value used in homework 6.8 (Description)}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
data(resfor6_8)
}
\description{
resfor6_8
The res value used in homework 6.8 (Description)
}
\keyword{datasets}
