% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resfor7Bchisq}
\alias{resfor7Bchisq}
\title{resfor7Bchisq
The res for 7B with chisq}
\format{
An object of class \code{data.frame} with 3 rows and 6 columns.
}
\usage{
data(resfor7Bchisq)
}
\description{
resfor7Bchisq
The res for 7B with chisq
}
\keyword{datasets}
