% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resfor7Bnorm}
\alias{resfor7Bnorm}
\title{resfor7Bnorm
The res for 7B with norm}
\format{
An object of class \code{data.frame} with 3 rows and 6 columns.
}
\usage{
data(resfor7Bnorm)
}
\description{
resfor7Bnorm
The res for 7B with norm
}
\keyword{datasets}
