\name{StatComp21047-package}
\alias{StatComp21047-package}
\docType{package}
\title{
  Homework created by chopong
}
\description{
  This is a package created by chopong which is also part of homework of Class Statistical Computation by Prof. Zhang Hong.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Chopong, <chopong@aliyun.com>.

Maintainer: Chopong <chopong@aliyun.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
