% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ts}
\alias{Ts}
\title{Ts
The Ts value used in CLiff walk (Description)}
\format{
An object of class \code{numeric} of length 10000.
}
\usage{
data(Ts)
}
\description{
Ts
The Ts value used in CLiff walk (Description)
}
\keyword{datasets}
